+++ src/background.js
@@ -1,4 +1,5 @@
 const URL_RULE_PATH = '/vuelos/resultados_ajax';
+const NOTIFICATION_ID = 'atrapalo-flights-';
 
 chrome.runtime.onInstalled.addListener(() => {
   chrome.declarativeContent.onPageChanged.removeRules(undefined, () => {
@@ -30,3 +31,30 @@ function sendMessage(message) {
     }
   });
 }
+
+function sendNotification(type, message) {
+  chrome.notifications.create(
+    NOTIFICATION_ID + type,
+    {
+      type: "basic",
+      title: "Atrapalo Flights",
+      message: message,
+      iconUrl: "icons/aeroplane_128.png"
+    }
+  );
+}
+
+function clearNotification(type, time = 1000) {
+  setTimeout(() => {
+    chrome.notifications.clear(NOTIFICATION_ID + type);
+  }, time);
+}
+
+chrome.runtime.onMessage.addListener((message) => {
+  switch (message.type) {
+    case 'COPY':
+      sendNotification('copy', message.payload + ' copied');
+      clearNotification('copy');
+      break;
+  }
+});