+++ CHANGELOG.md
@@ -2,10 +2,20 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.3.0] - 2017-10-21
+### Added
+- injected script in web page.
+- copy text functionality in result boxes, highlighted as blue text.
+- chrome notifications functionality.
+- notifications appeared when trigger copy function.
+- Price lines link in results boxes.
+
 ### Changed
-- remove webRequest dependency
 - change long-running extension (background persistent) to page Event.
-- fix hidden menu behaviour
+- fix hidden menu behaviour.
+
+### Removed
+- html title attribute info in results boxes.
+- webRequest dependency.
 
 ## [0.2.2] - 2017-10-15
 ### Changed